A bunch of updates all targeted at improving loading.
I switched the list view model away from using a listener towards a download approach assuming that the data isn't being updated often.

Firebase documentation doesn't express a method the allows all data to have a listener only the most recent data.

I simplified the get calls for sub collections as well.

I also updated the way the app knows if an initial download has occurred and that is through usingserDefaults, this was a big bug.

Now there is a persisted value to indicate if the app should attempt to load from .cache